refactor(annotate_types): revert Snowflake ROUND integer-to-DECIMAL conversion [CLAUDE]
BIGINT/INT are aliases for NUMBER(38,0) in Snowflake, so converting them
to DECIMAL(38,0) in the annotator output adds noise without adding
information. Revert the INTEGER_TYPES branch and restore the original
BIGINT/INT fixture expectations.

+++ tests/fixtures/optimizer/annotate_functions.sql
@@ -4427,11 +4427,11 @@ VARCHAR;
 
 # dialect: snowflake
 ROUND(42);
-DECIMAL(38, 0);
+INT;
 
 # dialect: snowflake
 ROUND(tbl.bigint_col, -1);
-DECIMAL(38, 0);
+BIGINT;
 
 # dialect: snowflake
 ROUND(tbl.double_col, 0, 'HALF_TO_EVEN');
